+++ src/main/java/faang/school/projectservice/dto/stageinvitation/StageInvitationFilterDto.java
@@ -1,4 +1,4 @@
-package faang.school.projectservice.dto.stage_invitation;
+package faang.school.projectservice.dto.stageinvitation;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
