+++ src/main/java/faang/school/projectservice/jpa/StageInvitationJpaRepository.java
@@ -2,7 +2,7 @@ package faang.school.projectservice.jpa;
 
 import faang.school.projectservice.model.TeamMember;
 import faang.school.projectservice.model.stage.Stage;
-import faang.school.projectservice.model.stage_invitation.StageInvitation;
+import faang.school.projectservice.model.stageinvitation.StageInvitation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
