+++ src/test/java/faang/school/projectservice/service/stageinvitation/StageInvitationServiceTest.java
@@ -1,24 +1,24 @@
-package faang.school.projectservice.service.stage_invitation;
+package faang.school.projectservice.service.stageinvitation;
 
-import faang.school.projectservice.dto.stage_invitation.StageInvitationDto;
-import faang.school.projectservice.dto.stage_invitation.StageInvitationFilterDto;
-import faang.school.projectservice.dto.stage_invitation.StageInvitationRejectDto;
+import faang.school.projectservice.dto.stageinvitation.StageInvitationDto;
+import faang.school.projectservice.dto.stageinvitation.StageInvitationFilterDto;
+import faang.school.projectservice.dto.stageinvitation.StageInvitationRejectDto;
 import faang.school.projectservice.exception.DataValidationException;
 import faang.school.projectservice.exception.EntityNotFoundException;
-import faang.school.projectservice.filter.stage_invitation_filter.StageInvitationDescriptionFilter;
-import faang.school.projectservice.filter.stage_invitation_filter.StageInvitationFilter;
-import faang.school.projectservice.filter.stage_invitation_filter.StageInvitationStageNameFilter;
+import faang.school.projectservice.filter.stageinvitationfilter.StageInvitationDescriptionFilter;
+import faang.school.projectservice.filter.stageinvitationfilter.StageInvitationFilter;
+import faang.school.projectservice.filter.stageinvitationfilter.StageInvitationStageNameFilter;
 import faang.school.projectservice.jpa.StageInvitationJpaRepository;
 import faang.school.projectservice.mapper.stageInvitation.StageInvitationMapper;
 import faang.school.projectservice.mapper.stageInvitation.StageInvitationMapperImpl;
 import faang.school.projectservice.model.TeamMember;
 import faang.school.projectservice.model.stage.Stage;
-import faang.school.projectservice.model.stage_invitation.StageInvitation;
-import faang.school.projectservice.model.stage_invitation.StageInvitationStatus;
+import faang.school.projectservice.model.stageinvitation.StageInvitation;
+import faang.school.projectservice.model.stageinvitation.StageInvitationStatus;
 import faang.school.projectservice.service.stage.StageService;
 import faang.school.projectservice.service.teammember.TeamMemberService;
-import faang.school.projectservice.validator.stage_invitation.StageInvitationValidator;
-import faang.school.projectservice.validator.team_member.TeamMemberValidator;
+import faang.school.projectservice.validator.stageinvitation.StageInvitationValidator;
+import faang.school.projectservice.validator.teammember.TeamMemberValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
