+++ src/test/java/faang/school/projectservice/validator/stageinvitation/StageInvitationValidatorTest.java
@@ -1,8 +1,8 @@
-package faang.school.projectservice.validator.stage_invitation;
+package faang.school.projectservice.validator.stageinvitation;
 
 import faang.school.projectservice.exception.DataValidationException;
 import faang.school.projectservice.model.TeamMember;
-import faang.school.projectservice.model.stage_invitation.StageInvitation;
+import faang.school.projectservice.model.stageinvitation.StageInvitation;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
