+++ src/test/java/faang/school/projectservice/validator/teammember/TeamMemberValidatorTest.java
@@ -1,10 +1,10 @@
-package faang.school.projectservice.validator.team_member;
+package faang.school.projectservice.validator.teammember;
 
 import faang.school.projectservice.exception.DataValidationException;
 import faang.school.projectservice.model.Project;
 import faang.school.projectservice.model.TeamMember;
 import faang.school.projectservice.model.stage.Stage;
-import faang.school.projectservice.model.stage_invitation.StageInvitation;
+import faang.school.projectservice.model.stageinvitation.StageInvitation;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
